pyproject.toml
[project]
name = "final-project"
version = "0.1.0"
description = ""
authors = [
    {name = "dayo98",email = "dayo98@naver.com"}
]
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.0",           # Django 추가
    "djangorestframework>=3.16.0",     # API 서버용 (선택)
    "python-dotenv>=1.2.1",           # 환경변수 관리용
    "django-cors-headers>=4.9.0",     # CORS 관리
    "django-allauth>=65.14.1",        # 소셜 로그인용 (allauth 추가)
    "PyJWT>=2.11.0",                  # JWT 처리를 위해 필요 (allauth 종속성)
    "boto3 (>=1.42.46,<2.0.0)",
    "firebase-admin (>=7.2.0,<8.0.0)",
    "paddlenlp (==3.0.0b4)",
    "transformers (>=4.45.0,<5.0.0)",
    "tokenizers (>=0.21.0,<0.22.0)",
    "pandas (>=2.2.0,<3.0.0)",
    "numpy (==1.26.4)",
    "torch (>=2.5.1,<3.0.0)",
    "torchvision (>=0.20.0,<1.0.0)",
    "torchaudio (>=2.5.1,<3.0.0)",
    "openai (==2.24.0)",
    "opencv-python-headless (==4.11.0.86)",
    "pillow (>=10.4.0,<13.0.0)",
    "einops (>=0.7.0)",
    "safetensors (>=0.4.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
